backend/requirements.txt
asgiref==3.6.0
boto3==1.28.21
botocore==1.31.22
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
setuptools==69.0.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.16
